==> 5_create_dataset_for_model_training.ipynb <==
import polars as pl

def compute_numerical_feature_from_categorical(dataset, y, categorical_feature_name, numerical_feature_name):

    feature_mean_value = dataset \
        .group_by(categorical_feature_name) \
            .agg(pl.col(y).mean()) \
        .sort(y)

    feature_mean_value = feature_mean_value.rename({y:numerical_feature_name})

    dataset = dataset.join(feature_mean_value, on=categorical_feature_name, how='left')
    dataset = dataset.drop(categorical_feature_name, numerical_feature_name+'_right')
    return dataset

### Read data

player_df = pl.read_csv('silver\Player.csv')
player_team_df = pl.read_csv('silver\PlayerTeam.csv')
team_df = pl.read_csv('silver\Team.csv')
team_league_df = pl.read_csv('silver\TeamLeague.csv')
nationality_df = pl.read_csv('silver/Nationality.csv')
foot_df = pl.read_csv('silver\Foot.csv')
position_df = pl.read_csv('silver\Position.csv')

### Join data

dataset = player_df.join(player_team_df, on='PlayerId', how='left')
dataset = dataset \
    .join(team_df, on='TeamId', how='left') \
    .join(team_league_df, on='TeamId', how='left') \
    .join(foot_df, on='FootId', how='left') \
    .join(position_df, on='PositionId', how='left') \
    .join(nationality_df, on='NationalityId', how='left') \
    .drop(['Team', 'Foot', 'Position', 'Nationality'])

dataset.head()

#### Save dataset to gold directory

dataset.write_csv('gold/Dataset.csv')
